requirements.txt
pytorch=1.9.1
torchvision=0.10.1
scikit-learn=1.0
jupyter=1.0.0
facenet-pytorch=2.5.2
matplotlib=3.4.3
black=21.9b0
cudatoolkit=10.2.89